feat: introduce tokio compact and remove AsyncOpShm
Given the strong coupling between `AsyncReadShm` & `AsyncWriteShm` and
`Stream`, and the need to explicitly call related `Stream` interfaces
(such as `close` and `reuse`) after using the interfaces of
AsyncOpShm, abstracting those traits is unnecessary.

However, for compatibility with `AsyncRead` and `AsyncWrite` of TokIO,
we added `StreamExt` to provide implementations for `AsyncRead` and
`AsyncWrite`. But since explicit calls to `Stream`'s interfaces are
still required in this case, we added detailed documentation to
`StreamExt` to remind users of this.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>